requirements.txt
Flask==2.3.0
pymongo==4.12.0
python-dotenv==1.0.0
